skillpath-ai/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pymupdf
pdfplumber
sentence-transformers
faiss-cpu
google-generativeai
httpx
python-dotenv
numpy
